contrarian-screener/pyproject.toml
[project]
name = "contrarian-screener"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "praw>=7.8.1",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "sqlite-utils>=3.39",
    "streamlit>=1.53.0",
    "typer>=0.21.1",
    "uvicorn>=0.40.0",
    "yfinance>=1.0",
]
